+++ cmd/project/project_test.go
@@ -1,0 +1,28 @@
+package project_test
+
+import (
+	"testing"
+
+	"github.com/benmatselby/hagen/cmd/project"
+	"github.com/benmatselby/hagen/pkg"
+	"github.com/golang/mock/gomock"
+)
+
+func TestNewRepoCommand(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+
+	client := pkg.NewMockProvider(ctrl)
+	cmd := project.NewProjectCommand(client)
+
+	use := "project"
+	short := "Project related commands"
+
+	if cmd.Use != use {
+		t.Fatalf("expected use: %s; got %s", use, cmd.Use)
+	}
+
+	if cmd.Short != short {
+		t.Fatalf("expected use: %s; got %s", short, cmd.Short)
+	}
+}