+++ pkg/provider.go
@@ -14,6 +14,8 @@ type Issue struct {
 // Provider is the interface to the back end data source
 type Provider interface {
 	ListIssues(query string, opts github.SearchOptions) (*github.IssuesSearchResult, error)
+	ListProjectsForOrg(orgName string, opts github.ProjectListOptions) ([]*github.Project, *github.Response, error)
+	ListProjectsForRepo(repoName string, opts github.ProjectListOptions) ([]*github.Project, *github.Response, error)
 	ListRepos(query string, opts github.SearchOptions) (*github.RepositoriesSearchResult, error)
 }