+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bumped docker image to Go 1.14 runtime.
 - Bump the build environment to test on 1.12, 1.13, and 1.14.
 - Provide a "--recursive" flag for `repo ls` so it will get all results for you. If not, it will ask you to "page" with the enter key.
+- Provide a `project` command to list projects.
 
 ## 2.0.0